requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.6.0
torch>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.18.0
python-dotenv>=1.0.0
openai>=1.3.0
gunicorn>=21.2.0
apscheduler>=3.10.0

